Update the invocation of Gameboard function
This way a new board state will be created when the Gameplay function.
If the user run the function or a function returning from it on console,
it will not be updated to the display. A new board state will be created
on the console without disturbing the display board state.

Earlier user could control the game using the console and it was updated
to the game display. Because I was updating the one object created using
the gameboard function. A new board was not created.